requirements.txt
bcrypt==4.0.1
certifi==2021.10.8
cffi==1.15.0
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.0.3
MarkupSafe==2.1.2
Pillow==9.4.0
pycparser==2.21
six==1.16.0
SQLAlchemy==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
